docs(inline-message): update alert docs and structure
