+++ .storybook/main.js
@@ -4,7 +4,8 @@ module.exports = {
     '../projects/canopy/src/lib/docs/principles/writing.stories.mdx',
     '../projects/canopy/src/lib/accordion/docs/guide.stories.mdx',
     '../projects/canopy/src/lib/accordion/docs/accordion.stories.ts',
-    '../projects/canopy/src/lib/alert/alert.stories.ts',
+    '../projects/canopy/src/lib/alert/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/alert/docs/alert.stories.ts',
     '../projects/canopy/src/lib/brand-icon/brand-icons.stories.ts',
     '../projects/canopy/src/lib/breadcrumb/docs/guide.stories.mdx',
     '../projects/canopy/src/lib/breadcrumb/docs/breadcrumb.stories.ts',
